+++ auth/presentation/src/main/kotlin/me/proton/core/auth/presentation/ui/signup/RecoverySMSFragment.kt
@@ -71,6 +71,6 @@ class RecoverySMSFragment : ProtonFragment(R.layout.fragment_recovery_sms) {
         viewModel.countryCallingCode.onSuccess {
             binding.callingCodeText.text =
                 String.format(getString(R.string.presentation_phone_calling_code_template), it)
-        }.launchIn(lifecycleScope)
+        }.launchIn(viewLifecycleOwner.lifecycleScope)
     }
 }