+++ auth/presentation/src/main/kotlin/me/proton/core/auth/presentation/ui/signup/TermsConditionsDialogFragment.kt
@@ -57,7 +57,7 @@ class TermsConditionsDialogFragment : ProtonDialogFragment(R.layout.fragment_ter
                     binding.root.errorSnack(R.string.auth_signup_no_connectivity)
                 }
             }
-        }.launchIn(lifecycleScope)
+        }.launchIn(viewLifecycleOwner.lifecycleScope)
         viewModel.watchNetwork()
     }