+++ user-settings/presentation/src/main/kotlin/me/proton/core/usersettings/presentation/ui/PasswordManagementFragment.kt
@@ -118,7 +118,7 @@ class PasswordManagementFragment : ProtonSecureFragment(R.layout.fragment_passwo
                     showSuccess()
                 }
             }.exhaustive
-        }.launchIn(lifecycleScope)
+        }.launchIn(viewLifecycleOwner.lifecycleScope)
     }
 
     private fun resetLoginPasswordInput() = with(binding) {